Assetmanager Spawning meteors and projectiles
instantiating meteors and projectiles now working

+++ SDLSpaceShooter/SDLSpaceShooter/TransformComponent.h
@@ -38,7 +38,7 @@ struct TransformComponent : public Component
 		scale = sc;
 	}
 
-	TransformComponent(float x, float y, int h, int w, int sc)
+	TransformComponent(float x, float y, int h, int w, float sc)
 	{
 		position.x = x;
 		position.y = y;
